+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Created TODO.md inspired by the 'Keep a Changelog' format
+- Add quiet setting and flag to suppress regular output
 
 ## [0.0.2] - 2022-08-28